Renamed & Organised QuizCard by sam was renamed to FlippingCard
