+++ frontend/src/components/FlippingCard/FlippingCard.css
@@ -1,0 +1,43 @@
+.flipcard-container {
+  /* height: 20rem;
+    width: 30rem; */
+  height: fit-content;
+  width: fit-content;
+
+  border-radius: 1rem;
+  box-shadow: 0px 10px 8px #999;
+
+  display: flex;
+  justify-content: flex-end;
+  align-items: flex-end;
+  /* padding: 2rem; */
+  margin: 0.5rem;
+}
+
+.front {
+  /* height: 20rem;
+    width: 30rem; */
+  height: fit-content;
+  width: fit-content;
+
+  /* border-radius: 1rem;
+    box-shadow: 0px 10px 8px #999;
+    display: flex;
+    flex-direction: row;
+    background-image: linear-gradient(rgb(44, 0, 109), black, rgb(44, 0, 109));
+    height: fit-content;
+    width: fit-content; */
+}
+
+.graph {
+  /* height: 20rem;
+    width: 30rem; */
+  height: 392px;
+  width: 400px;
+  /* height: fit-content;
+    width: fit-content; */
+
+  border-radius: 1rem;
+  /* box-shadow: 0px 10px 8px #999; */
+  display: flex;
+}
