+++ frontend/src/screens/QuizPage/QuizPage.jsx
@@ -2,10 +2,11 @@ import React from "react";
 import { AdminHero } from "../../components/AdminHero/AdminHero";
 import "./QuizPage.css";
 import { useParams } from "react-router-dom";
-import { QuizCard } from "../../components/QuizCard";
-import { QuestionCard } from "../../components/QuestionCard/QuestionCard";
+import { FlippingCard } from "../../components/FlippingCard/FlippingCard";
+import { QuestionCard } from "../../components/QuestionCardRuben/QuestionCard";
 
 const QuizPage = ({ quiz }) => {
+  //! NOTE: this list is for rubens question card
   const questions = [
     {
       question: "Capital Of India?",
@@ -45,13 +46,12 @@ const QuizPage = ({ quiz }) => {
         triggerHandler={triggerHandler}
         clickHandler={clickHandler}
       />
-      <QuizCard
+      <FlippingCard
         title="RoboWars"
         description="Nov 15, 6:00PM"
         imgSrc="https://picsum.photos/330/320"
         percentages={[0.1, 0.3]}
       />
-      <QuestionCard questions={questions} />
     </div>
   );
 };
